+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update Vulkan Headers to version 212
 - Change from `Box` to reference in `create_{instance,device}_fn`
+- Pointer chain wrapper args are now a mut ref
 
 ## [0.22.0+204] - 2022-01-25